Rescue on error when update to User's Time zone fails
Still seeing the `Etc/Unknown` Time Zone string coming from our JS
script in production. This causes Rails to choke when attempting to set
`Time.zone` for the current User session. Since we really can't help
this any better way than defaulting to the application's Time Zone, we
just swallow the error and move on. There may be a way to detect if the
Time Zone is valid instead of using error handling to deal with this,
but I'm OK with this approach since it is a non-standard flow / somewhat
exceptional situation.