Fix a few issues with site meta tags for SEO
- Add missing page title on About page
- Add unique page descriptions for About, Metrics, Users Show, Games
- Fix description meta_tag across site to be the shortened version of
  the site description found in `site.meta_description` in en.yml
- Set a canonical URL for the hundreds of Games Show pages to point back
  to the main Games Index page for SEO collection
- Change the Open Graph (OG) image to be a smaller, PNG image. Hoping
  this will fix missing images in shared links.

TODO:
Still not sure why DuckDuckGo is listing the site/pages with a "title"
(or something?) of "minesweeperalliance.net"... e.g.:

```
minesweeperalliance.net    # <--- "Title?" Or what??
minesweeperalliance.net
War Room — Minesweeper Alliance
Welcome to Minesweeper Alliance, an open-multiplayer riff on the classic Minesweeper game. Join forces with your fellow minesweepers to take on the mines in real-time! But be warned … We've got our work cut out for us! The mines are randomly placed, and the game is not won without clear strategy
```
... other sites don't seem to have this problem...