Fix spacing issue in Just-ended Game Results view
We were getting an unexpected margin-top applied to the Duty Roster
heading on the just-ended Game results view... the is was caused by the
combination of the `space-y-12` on the container element (intended) and
the fact that `turbo-cable-stream-source` elements are apparently a
block-level element (wtf?). So, to prevent the
`turbo-cable-stream-source` element form playing a part in our spacing
scheme, we have to wrap it into a neutral div. Ew.