Switch all tooltips over to new Floating-UI implementation
The new Floating-UI implementation is much better because it takes care
of keeping tooltips on-screen as much as possible, even when scrolling
or resizing the window. So, we're getting rid of our CSS-only solution
in favor of Floating-UI (see the previous commit for initial setup of
this library + CSS).

To implement the changeover: we add an application/_tooltip.html.erb
partial to serve as the common view template for rendering tooltips. By
using strict locals, we achieve an "object-lite" interface that keeps
ease-of-use in mind.
- i.e. I briefly started down the road of using View Models to store
  tooltip style/config attributes, but this was too heavy handed for
  what it is. I immediately started wanting more of a config-style
  interface for it. Which then got me thinking that, really, the view is
  what defines the config here. Basically a utility CSS-style approach,
  but for tooltip code. That said... we can DRY things up still. Thus,
  the _tooltip partial!

Note:
The primary pattern I'm anticipating here for rendering a tooltip's
description text is to reference a "term" from our en.yml translations
file. This can be overridden by passing in a `description` local
variable/value.

Also:
- Remove our old, CSS-only solution to tooltips.
- Update tooltip terms a little bit, and relocate them within the
  en.yml file so that they don't cut through the middle of our
  page-related translations.
- Remove the `cursor-pointer` style from `dfn` elements by default.
  That change was originally motivated by a need we no longer have:
  toggling the Fleet Roster via a click onto the Fleet Count text, which
  I was styling in the same style as `dfn` elements. These probably
  should not have been combined at the time anyway. But, alas, I do
  think keeping the default cursor style for `dfn` elements is better as
  it visually allows e.g. text selection, which is good.