Update "participation status" in fleet roster on Cell Actions
From the last commit:

We now render
- ⛴️ next to active participants, and
- ⚓️ next to passive observers.
This is based on whether or not the User has any associated
CellTransactions for the current Game.

In this commit:

We will now update the current participation status per each User
without having to re-render the fleet roster partial, and, thus, without
having to re-query User status.

We do this by using the FleetTracker registry/cache to track current
participation status by User#token.

We then update the icon (only) when the FleetTracker confirms a change
from passive -> active sate for the current User (token).