Don't render board updates to originator, just broadcast
Up until now, the best we've had was just to rely on the
versioned-replace logic in the Turbo broadcast to not re-render the
Game board after a rendering the result of a Cell Action.

This seemed better because it's faster / more immediate.

But, we're going to try just relying on the broadcast for a while to
see if that's not fast enough.

Why?
- Less data to send over the wire. We can just respond with `head (:ok)`
  for the originator request. And the broadcast was going to happen
  either way.
- Logically easier to understand.
- This is one step towards my next effort: Update just the Cells that
  changed, not the whole Game board.
  - If we still have to render a turbo_stream AND broadcast updates,
    we'll be doubling up on that logic.