Fix unneeded passing of Game object to Cell view model
Passing Game on to Current Game Cells is a vestige from months ago at
this point. We do this for Past Games, but what used to make us mirror
this requirement in Current Game Cells is no longer around. So we can
remove this old code as it was just confusing.