Use `dom_id` for outputting Cell IDs in Game Board view
TIL that DOM IDs must start with a letter, not a number. Else, they're
unqueryable with JS. e.g.

```javascript
// Will fail:
document.querySelectorAll("#139873, #139874")

// Works:
document.querySelectorAll("#cell_139873, #cell_139874")
```

In light of this, we need to update our Cell IDs in the Game Board view
to use the `dom_id` helper, which outputs e.g. `cell_139873`. Then we
just have to adjust the board_controller.js to extract the Cell ID from
this string for dispatching a request based on it.