Broadcast direct Cell-specific CSS replacement on DehighlightNeighbors
Instead of re-rendering the entire board on a Cell::DehighlightNeighbors
action, we will use a custom TurboStream action + JS handler to directly
update the CSS classes for the specifically applicable Cell objects.
i.e.

On Cell Dehighlight:
1. Call Cell#dehighlightNeighbors and store off the updated Cells result
2. Combine:
  - Build a custom Turbo Stream tag with:
    - custom action "replace_css"
    - a `target` attribute set the originating Cell (by DOM ID)
    - a `css` attribute that contains the relevant CSS classes for
      displaying revealed Cells in the view, as determined by appealing
      to Games::Current::Board::Cell.
  - Build a set of custom Turbo Stream tags with:
    - custom action "replace_css"
    - a `target` attribute set the updated Cell (by DOM ID) from
      number 1
    - a `css` attribute that contains the relevant CSS classes for
      displaying non-highlighted Cells in the view, as determined for
      each Cell by appealing to Games::Current::Board::Cell.
3. Our custom TurboStream JS handler picks this up when the HTML
   elements attach to the DOM and then performs the CSS styles
   replacement on each of the found Cells in the DOM.
   - Note: This is a full CSS class replacement as Cells are entirely
     styled by their current state.

NOTE: For De-highlight, we have to render get the CSS for each of the
updated Cells individually since the resting state for a Cell can vary
even just within the set of updated Cells (that we just dehighlighted).
- This mainly only matters when viewing the Game in Debug mode right
  now... but I'm keeping it open for other reasons in the future.
  (Hoping it's worth it! Though the cost isn't large at all... it's
  just not as cool!)

This has the distinct advantage of not requiring a board rebuild to get
done. This is more performant, but it also preserves mouse events, etc,
for the performing user plus all other users (potentially).