Fix: Repeated Flag toggles stop animating on mobile
Previously, the first Flag that was added would apply the CSS class
"animate-ping-once" and then any subsequent events would attempt to do
the same to no effect. Instead, we need to remove the class after the
animation is done so that subsequent Flag toggles have the opportunity
to start a new animation (i.e. indication of success).

This wasn't a problem when the feature was first implemented because
the board was completely replaced on Flag toggle back then. Now, with
our recent changes to only update the affected Cells on Flag toggle, we
have to tidy up the DOM element's state ourselves.

Note: I'm itching to move this `animateOnce` function out into a lib of
some sort, but I'm just not sure what yet. So leaving it for now as
we'll know more later.