Fix short/longPress events on mobile
This fixes longPress vs tap (short press / touch) detection on mobile.
Previously we didn't need to do more because we could rely on the Game
board being replaced on cell actions / submission. But we're not always
replacing the Game board anymore (and trending towards "never"). So this
commit makes the game playable again on mobile.

Note: Previously we relied on the standard `click` event to do Cell
Reveals for us. But, now that the Board isn't reloading on us anymore,
it's actually best to not listen for `click` on mobile anymore because
it will just always fire--even after a longPress event. So we have to
detect Short Press and dispatch Cell Reveal ourselves.