Restore "RevealNeighbors" on mobile
The final link in the chain after removing `click` for mobile (due to
conflicts now that the Game board doesn't get replaced on Cell Actions--
see previous commits) is to restore "RevealNeighbors" functionality.

This required bypassing the standard `revealNeighbors` function, which
tries to classify the `event` as a proper "right click"-like event,
which doesn't pass on mobile. So now we have an internal
`#revealNeighbors` function that is the "Do it!" version that doesn't
process the event type first.
- We also separate out the touch-related code into distinctly named
  `#tap` and `#longPress` functions while we're at it, to follow suit.