board_controller.js -> desktop & mobile-specific controllers
board_controller.js was getting way too large, and it was getting hard
to differentiate mobile vs desktop functionality. Which has suddenly
become important!

Also, take advantage of some namespacing to group things together... and
to mirror what we're doing in the views and controllers, for
consistency.

All functionality remains the same after this refactor.
- The refactor touched a few other JS files as well, to keep them up to
  date with my latest understanding of how exports work, etc.