Use Cell::TurboStream::Morph for all Cell updates!
This commit marks the end of the need to refresh then entire Game board
for every Cell action. Instead, we now just update only the Cells that
have been updated/changed. Which we do using our new-found
TurboStream #replace with method `:morph`! (See my previous commit.)

To make this possible:

Ensure that Cell::Reveal and Cell::RevealNeighbors keep track of the
total set of Cells that they're making actual changes to. Then expose
that set to the calling context. e.g.:

```
 Cell::Reveal.(current_context).updated_cells

 # OR

 Cell::RevealNeighbors.(current_context).updated_cells
```

And to make *this* possible:
1. Return `nil` if no change was made/needed on:
  - Cell#reveal
  - Cell#soft_reveal
2. Ensure that Cell::RecursiveReveal can handle `nil` actions properly,
   and doesn't include `nil`s into the set of "revealed" Cells that it
   then returns

Also:
- While working on Cell::RecursiveReveal, I decided it would be easy
  enough to turn this into a module instead of a Class that we must
  instantiate. This should further help reduce the call stack, which
  is good since it's recursion!

Note: This commit leaves us in a state where we don't see Game-end...
since that takes a Game container refresh. That's coming in the next
commit.