Remove our custom "versioned_replace" TurboStream Action
Now that we're doing broadcasts for everything, we don't need versioned
replace of the Game board anymore. The versions have now sub-divided
themselves greatly... into just Cells! That is, aside from the
Whole-Game-board replace that occurs on Game end. Which doesn't need to
be versioned.

This commit is an intermediate step on the way to using 100% broadcasts.
See the next commit!