Update "placed Flags count" on the ToggleFlag Cell Action
This is yet another thing that used to be handled only by replacing the
entire Game board on Cell Actions. Now that we're no longer doing this,
we have to manage our own state for the "placed Flags count" in the
Game board header.

To support this, we create a way to pass additional updates via a block
on the `broadcast_updates` method.

I chose to prepend the block somewhat arbitrarily... but it made it
easier to see that it was working from the Rails server output logs.
e.g.

[ActionCable] Broadcasting to current_game: "<turbo-stream action=\"update\" target=\"placedFlagsCount\"><template>41</template></turbo-stream><turbo-stream method=\"morph\" action=\"replace\" target=\"cell_145191\"><template><!-- BEGIN app/views/games/current/board/_cell.html.erb -->\n<td\n  id=\"cell_145191\"\n  data-revealed=\"false\"\n...

Note: we need not doing anything different than just the `join` that all
of this is feeding into in WarRoomChannel.broadcast, even though a
missing block results in a `nil`, and even though we can now have a
nested Array in ... because `join` is awesome:

```
[1, nil, [2, 3]].join()
=> "123"
```