Fix "Reveal Random" button to work in mobile
I guess this has been broken ever since we got rid of `click` events
detection. Now, to get mobile working again, we have to manually
dispatch a `touchend` event to properly simulate a tap on a Cell.

Note: we must use `bubble: true` so that the event can bubble up to the
`<tbody>` on which our stimulus controller resides / is listening from.