Remove Game::Current view model info Games::Current::Status
The only thing we were using Games::Current for was status-related
methods. So that fits better in Games::Current::Status anyway.

Plus, in an upcoming commit, we'll remove the
Games::Current::Status.broadcast_status_update class method entirely,
which was a big part of the driver for this separation in the past--
though even then, I don't think it was the right call anymore.