Handle now-needed view updates on Game start
Now that we're no longer refreshing the entire board contents on every
Cell Action, we have to handle each necessary view update as one-offs.
There are a few special cases like this on Game start:

- Update Game Status from "Standing By" to "Sweep in Progress"
- Start the Game Timer
- IN DEBUG MODE: Reveal mine cell indicators

We pull all of these updates together into one place, within
Games::Current::Board::Cells::RevealsController. I think this is a good
job for a Controller to take on.
Notes:
- Previously, the broadcast for refreshing the Game status was firing
  far too often--on every Cell Reveal--due to a false assumption based
  on the method name it was added to: `start_game_if_standing_by`. Even
  though I've moved this process out of that method, I've gone ahead
  and added the proper guard clause on that method to help it live up
  to its name in the future.
- The other advantage of pooling all these updates together is that they
  can then all be sent in the same broadcast. At least... I assume
  that's an advantage.

Also, to support this, we have to add the concept of Game#just_started?.
We do this the same way we did Game#just_ended? -- with a virtual
variable(attr_accessor) that we set directly on the Game instance
(without storing it) for use during the current request cycle.