Remove the "Jump to Current Game" btn from page nav
I'm appreciating the firmer line between the War Room and the
Past Games / Archive. So, just to de-complicate things a bit more, we
remove the "War Room" button (the circle with the right-arrow in it)
from Past Game views.

Also remove the old "circle with right-arrow" icon as well, as I forgot
to do this when I remove the similar "Jump to Most Recent Past Game"
button from the War Room's page nav.