`current_game` -> `game` (as much as is sensible)
I've been slowly making this transition over time, where new View Models
will refer to its Game as just `game`, even if we have to pass it in as
`current_game` from the caller.

This commit finishes out that transition. We prefer just `game` because
that's what the model is. However "Current Game" is still valid as a
disambiguator in certain places. e.g. on the Home page where we're
trying to distinguish between an active Game being present or not, and
what to do otherwise (which is: show the "New Game" view).