"New Game Notification" -> "Fleet Mustering Notification"
This changes the back-end and front-end messaging to be a bit more
descriptive and on-theme. I actually think this makes more sense to me
now even on the back end. That said "mustering" (or "muster") might be
a new term to many.

Muster: verb [with object]
1. assemble (troops), especially for inspection or in preparation for
   battle: "17,000 men had been mustered on Haldon Hill."
  - [no object] (of troops) come together for inspection or preparation:
    "the cavalrymen mustered beside the other regiments."
  - [no object] (of a group of people) gather together: "reporters
    mustered outside her house."

Also, as a notification, "Fleet mustering!" feels more accurate than
"New fleet deployed!" since the battle hasn't actually, necessarily
started yet.