Core rewrite of FleetTracker to fix numerous issues
FleetTracker updates were all kinds of messed up. This commit makes the
following behavior and/or reliability changes:
- Don't remove a user from the Roster until their 2-second expiration
  period has past
- Don't send unneeded Roster list updates (ensure the User isn't already
  in the Roster first)
- Don't send unneeded Fleet count updates
- Don't send unneeded User Participation Status updates
- Once a User is marked as an active Participant, they're always an
  active Participant now (until the next reset -- on new Game)
- Keep Users sorted by join time regardless of whether the Roster was
  loaded from initial page load, a refreshed page load, or just User
  additions/removals/re-additions from Turbo stream updates.
- Otherwise just push Registry-specific functionality down into
  Registry. We were violating encapsulation before.

Of note: It was fun to get rid of the "primitive obsession" code smell
by converting FleetTracker::Registry entries from Hashes to actual
FleetTracker::Registry::Entry objects. Way better!

There is still the issue of WarRoomChannel subscribe and unsubscribe
events coming in out of order... but I don't know what to do about that
one. I guess we could go nuclear and look for e.g. window events and
use that as an indicator of a User's presence to inspire a Roster
re-add if missing.