Separate out Game <-> User relationships into 3 types
I think of this as:
- Users connected to a Game in any way are "participants" (in general).
- Users that actually contributed to a Game through any Cell Action are
  "active participants"
- Users that joined a Game and just watched, never contributing, are
  "passive participants", or "observers".

So, in summary: Games -> Users and Users -> Games can be:
- "observed (only)"
- "actively participated in"
- "all" / "any"

Or, in specific:

User#games =
  Games that were joined in on by this User--in any fashion.
  (Whether or not they were actively participated in by this User).
User#observed_games =
  Games that were joined in on, but never actively participated in by
  this User.
User#actively_participated_in_games =
  Games that were actively participated in by this User.

And their reciprocal:

Game#users =
  Users that joined in on this Game--in any fashion.
  (Whether or not they were active participants in this Game).
Game#observers =
  Users that joined in on, but never actively participated in this Game.
Game#active_participants =
  Users that actively participated in this Game.

This is perhaps a bit odd, but the distinction is actually important,
and I'm finally getting to spell it out concretely here.

There was only one place where we ended up preferring the "all"
relationship, so far (not caring about "observers" vs
"active participants"): for the cache key on Past Game Show pages. Since
we list both observers and active participants on these pages, we need
to bust the cache if any have changed their username since the
fragment was cached. So participation status really doesn't matter.

And more to come from this...