Remove DB User sorting by joined_at & participated_at
I just can't. These queries are far too complex and un-extensible.
Something has to change about the design if we're going to sort Users
by when the joined a Game vs when they first participated in a Game.

I'd rather do it in pure Ruby at this point under the current design.

For now, in order to keep moving forward with what I actually wanted to
be working on, I'm just tearing this all out and leaving it for another
time. Users will be unsorted at this point. Search on `FIXME` to
re-insert sorting at a later date.