Fix: Properly recover from failed Cell Actions
I had intended `safe_perform_game_action` to capture errors and then,
if one is encountered:
1. halt processing
2. Refresh the user's page to get them back on track

However, neither of these were actually happening since rendering
doesn't halt processing.

To fix: We add `return if performed?` after each of the
`safe_perform_game_action` blocks. This is a bit of a bummer to have to
do, so I will think about a cleaner solution in the future.