Add FlatArray and use it for concatenating updated Cells
I didn't like the interface of defining/using a `push` method on the
service objects... but wasn't sure what better to do. But, that was just
the primitive obsession code smell again.... which led me to creating a
new FlatAray class.
- Basically `push(...)` isn't as self illuminating as
  `updated_cells << ...`.

FlatArray is an Array that presents a friendly interface for ensuring
single or multiple-object pushes are always nicely concatenated on,
instead of creating an Array of Arrays. It should otherwise act like an
Array, as needed.