Create ParticipationTransactions on toggling a Flag
Previously, we weren't recording a Flag toggle event as a Participation
event, when it should be. This created a disconnect between what the
Fleet Roster showed as participation and what the Game itself showed
on page refresh. So we'd see:

1. Toggle a Flag
2. Which causes the ⛴️ to appear next to the user's name in the Roster
3. Refresh the page and the ⛴️ goes away
4. Do any Cell action and the ⛴️ doesn't come back since the
   FleetTracker thinks of the User as having already been activated.

So basically we just need to keep the database in alignment here.

Note: we will still have the same problem with HighlightNeighbors...
but I'm OK with that for now because I plan to try to remove
HighlightNeighbors from the database altogether soon. Will come back to
this if those plans fall through, though.