Remove cells.highlighted from database; Use virtual attr instead
Now that we don't actually refresh the entire board on Cell Actions, we
don't need to store a Cell's "highlighted" state in the database.
Instead, we can just send the styling for it over the wire (broadcast)
and all listening clients will see the effect. But, it need not survive
page refreshes.

To achieve this, we can just define a virtual attribute on Cell and use
that to hold this temporary state--just long enough to build the
appropriate HTML to send over the wire.

This is actually an improvement on what we had before because:
1. Fewer database hits! And this is actually somewhat of a big deal
   since this hit was occurring on every "Reveal Neighbors" event.
2. It somewhat avoids "bugged-out" situations where e.g. if a user
   clicks to "highlight" neighbors--but then drags their mouse away (so
   that the de-highlight can't occur)--the highlight effect would get
   stuck! Lasting until manually cleared by a good samaritan--even
   persisting beyond page refreshes. Now, however, in such a situation,
   a simple page refresh can cure these visual artifacts.