Identify ActionCable streams by the actual current User
I didn't realize we had this level of control before from within the
ApplicationCable::Connection object... but this is cool.

We also add a logger tag to display User#identifier while in the
development environment. I had figured it out earlier that there was one
broadcast event per active websocket connection, but this spells it out
and show which User each is going to, which is cool.

The work of adding tags to the logger doesn't fire on every websocket
broadcast, but rather just once, when the connection is first
established. They are sticky per session from then on.

Reference: https://toxigon.com/websockets-in-rails