Turn off Turbo prefetch for most links
Instead of zeroing in on each link or link type to not pre-fetch, this
commit just disabled it for entire pages.

See: https://turbo.hotwired.dev/handbook/drive#prefetching-links-on-hover

I'm making this change, not because I've seen and measured that we're
seeing too much activity on the site. But, just out of a gut feel that
it's not worth the extra requests when hovering over links. We've
already got so many requests coming in just for the Game play itself!
So I'm just proactively making this change. Besides, the pages load so
fast on their own that the pre-fetch doesn't seem to add much of
anything.