Fix: Game Links on Past Games index pg too wide for content
The direct ancestors of a `grid` are "grid-items", which are block-level
elements. So we have to insulate the links with a <div> so that the
links themselves can escape being the full width of each column.

What makes this weird is that we want links to flow left-to-right and
*then* top-to-bottom. Instead of a normal grid flow where everything in
the left-most column is defined first (top-to-bottom) and then the same
for the 2nd column and so on.