Refactor "Profile" -> Current User / "Account"
Today.. this feels more accurate, as it's a page for "account"
management type stuff. More-so, at least, than profile type stuff.
And, we nest it below CurrentUser so we have a common base for additions
in the future (like shared sub-section nav), plus an obvious indicator
that the "Account" (etc) is in reference to the "current user".

I'm thinking that a "profile" is more like the way that the player is
present within the game. Like their name. There isn't currently an
option to change one's name outside of the Just-Ended Game screen, but
if we do add it, then a Profile page/tab will make sense.