Tidy: Clean up the Welcome Banner code a bit
This commit is extracted from upcoming work to allow for Guest Users.
There are no functionality changes, just shuffling of responsibilities
to match latest understanding/design.

The biggest change is that we're no longer operating under the
assumption that an object should define its own Context wrapper.
Instead, we should always pass in a proper Context to any object that
uses one. This follows the normal rules of OOD--pass in duck, don't
pass in an object and make it act like a duck therein.
- As a direct benefit, multiple contexts can then pass in their own,
  different reasons why something within the context should be as it is.
  e.g. when deciding whether or not to show the WelcomeBanner. We'll
  make use of that later, though, as this is just the code shuffling
  prep work.

Also:
- Remove the turbo_frame_tag for the welcome banner, in favor of just a
  div with an `id` set (turbo_update_id), as is our latest convention.
  The idea being: only use `turbo_frame_tag` when there's an actual
  need for a frame. Else just use a light-weight div.