Send Cell Action updates back via the response as well
Instead of just relying on the Broadcasts everyone else receives.

This helps, to some degree... maybe, in situations where the websocket
connection has simply gone away. i.e. no Broadcasts are received.
However, the user can still click the game board, which sends requests
to the server and affects the Game board state. However, without the
request response, the user can't see that Game board state change...
It doesn't solve the full problem though because who knows what else is
actually supposed to be visible on the Game board that the current User
just can't see. In fact, if they click on suck spaces, what comes back
is a "paint"-like even on that cell -- a revealing of the state that
Cell should have been rendered as, and *is* stored as on the server.

This code was extracted/saved from my `sourced_replace-dead_end` branch.