Move `.largest_id` out of ApplicationRecord into Game
This method is only being used by Game and makes no sense in the context
of e.g. a record that uses a UUID for its ID.

Also, start caching the result onto the class. This will mean needing to
do a server restart to pickup changes, but they won't happen often
enough to warrant constant recalculation.