Build a Guest i/o creating Users, until active Game participation
This commit solves the problem of having a bunch of invalid Users in
the database just from random visits to the site. e.g. when web
crawlers/bots visit, we would get a User created for no good reason.

So, instead of just always creating a User object (+ DB record), we will
now use a Guest as a stand-in, up until active Game participation
occurs. At that moment, a User record will be created.

We also must expand the definition of active Game participation to
include Game creation for this to make sense. i.e. we require a User to
attribute the Game creation event to.

Guests are transparent to the FleetTracker / Home Roster. I started down
the path of showing Guests in there, but it's far too much state to
keep track of... and ultimately doesn't even work after a page refresh.
On page load/refresh, the server handles rendering the roster and the
FleetTracker only affects updates. But if there is no entry in the
database for a Guest then there's no way to render them in the Roster.
So it just doesn't make sense.