Fix: Don't show "Custom" for non-signers on Just Ended Game view
This was ultimately caused by the choice to both render via response AND
(again, for the originating user) via broadcast ... while using method:
`:morph`. We can keep the former if we remove the latter.
- Morphing causes our stimulus controller to not be called on the 2nd
  rendering (the broadcast), which therefore fails to remove the
  "Custom" button, as expected.

Also, fix unreliability of the `"unsignedUserNav"` target in the
app/views/current_user/_nav.html.erb partial now that we have Guest
users and, therefore, changed this branch of partial to render for all
participants instead of just for signers. So, now, better to just
output a true/false data attribute for signer status to all in the
`participant?` branch of the partial.

Also, Switch to newer naming style turbo targets:
`turbo_update_target` -> `turbo_target` for the user nav partial.