Remove DEV_MODE and DISABLE_TURBO app flags
Disabling turbo just isn't realistic anymore. And for anywhere it's
maybe needed some day, we can just do it manually. Being so obscure,
it's likely one wouldn't think to go turn it on anyway.

Dev Mode was useful at the time, but only for one thing: disabling
validations on mine density for Custom games. This can easily be done
as a one-off and also suffers from obscurity.

So, just alleviating the burden on the code seems best in both cases.