Stylize User Show page as a "Personnel File"
Now that we have called the User Show page the "Personnel File"--in the
User Account Page, it makes sense to use that as a styling hint. So, we
add a manila folder-like top tab for the user's name. This adds a little
something over just the underline we had before.