Fix: JS error about null $meta_tag on server errors
No User should ever see this, but while debugging errors in development,
I noticed this JS error--which was a red herring--so wanted to fix it.

Basically: we won't always be able to access the
`meta[name="theme-color"]` meta tag so... abort gracefully when that's
the case. (Essentially by just doing a nil check first.)