Show "Personal Best!" per Metric Type per User on Game end
It's nice to be able to show active participants when their performance
metrics are their personal bests for that metric type. e.g.
- if a user was an active participant
- and they achieved their best-ever Game Score then show that
- or if they achieved their best-ever Game 3BV/s then show that
- or if they achieved their best-ever Game Efficiency then show that

We show Personal Bests with a `🥇` icon and by highlighting the text in
a gold/amber color.

Since this is a bit of an expensive operation, we try to only perform
these queries and build the requisite objects if we have to and memoize
whatever intermediary objects we can.

We're able to render *Personal* Bests because we're using a lazy-loaded
Turbo frame to render
`games/just_ended/<active_participants|observer>/footer` for each
individual viewer ({User}).

However, for now I have chosen to only apply the Personal Best
indicators to the Just-Ended Game view, and not the Past Game view so
that we can cache the Past Game view... vs recalculating Personal Bests
per User. We may be able to get away with caching the Past Game view per
User... but that's obviously a lot less efficient.

To support this split between Just-Ended and Past Game views, I've had
to break out a separate Just-Ended version of the Metrics view model +
partial as well as our newly named sub-category: PerformanceMetrics.
(The non-"Performance Metrics" would be called "Activity Metrics",
though I didn't have to formally name those yet.)

Also, add a "Colors" section to UI Portal Home page. Used this to figure
out what is the best Gold-like color to use in light/dark mode for
representing personal best text highlights.