Explicitly require useragent gem to fix occasional failures
UserAgent is already a dependency of actionpack, but isn't already
loaded when needed. Typically, it's only when one calls the
`allow_browser` macro that `require "useragent"` gets called... and, so,
our use case is abnormal.

Instead of just calling `rquire "useragent"` somewhere, it's better to
call out our dependency by adding it as an explicit gem in the Gemfile,
which will also handle requiring it for us.