Add a DEBUG-mode auto-solver to aid development
There's now a pencil icon in the dev_flags overlay at bottom-right that
will toggle (play/pause) a random auto-solve on the board.
- This only appears when the DEBUG=1 flag is set when starting the
  server as it needs the darker squares to be present that represent
  mines in order to work successfully. Else it'd just be clicking in
  the dark, not knowing where the mines are.

To support the new button, I shrank the UI Portal link down to just be
on the breakpoint indicator text instead of the entire dev flags
overlay.

Also:
- Move the overlay to the left side of the window instead of the
  right side as it was getting in the way of the theme switcher while
  there's not really anything going on on the left side.
- Convert `div`s -> `span`s, as only makes sense, semantically.