Update to gemwork v0.7.7 & remove custom brakeman rake task
Moved this rake task up to gemwork for all projects to benefit.

Also update other gems...

Note: The new brakeman task in gemwork fixes an issue where it wasn't
showing issues if any exist. Found this issue in another project, but
basically: while it says there is a warning present it doesn't display
details on it, which makes it hard to notice.