Fix tooltip alignment/placement issues
Previously, the arrow would appear to be not quite in the center of the
tooltip element. The big problem was really just the margin styling on
the main tooltip element. Oops.

Note: To figure this out, I briefly commented out the `hide()` portion
of the JS in app/javascript/controllers/tooltip_controller.js.

After doing this, I also had an opportunity to perfect the arrow size
and position.

Also: Add padding to the `shift` middelware component. This prevents the
tooltip from running into the side of the window where relevant.
- Moving the `shift` middelware down below `flip` probably didn't do
  anything... but I did so based on vague language from the docs.
  https://floating-ui.com/docs/shift