Broadcast new Bests on Game end!
On Game end, we will now determine if the Game was a record (a.k.a a
"Best") of some kind. If so, we'll broadcast a flash notification
indicating as much.

We use a slightly randomized flash message for this to add a little
spice. Especially for situations where more than one Best is achieved
at once.

Flash Notifications are broadcast to everyone across the whole site 😎